librados: return int64_t pool ids
468e28ee60ee2fe625d2680c792a4bcb9ef19951 missed the get_id() functions.

+++ src/include/rados/librados.hpp
@@ -324,7 +324,7 @@ namespace librados
 
     void locator_set_key(const std::string& key);
 
-    int get_id();
+    int64_t get_id();
 
     CephContext *cct();
